requirements.txt
flask==1.1.2
joblib
scikit-learn==0.24.1
dill
emoji
streamlit
pillow